move func in main.c
