+++ inc/strop/string.h
@@ -9,4 +9,6 @@ int str_len(const char *str);
 
 int str_index(char source[], char searchfor[]);
 
+int getline_(char line[], int max);
+
 #endif
